Implement pseudo collection using config data
By nature of its interface, dataproviders were made to working with a
collection. What we did here is remove the need for collections, by
"reimplementing" our own way of sorting and paginating.

We also implemented a cache to save records because when a user sorts,
or paginates, it will call the getData() method every time. This will
make the component feel sluggish because we need to call a double nested
loop every time we want to sort or paginate.